+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.2.0
+
+- Add VM destroy command
+- Add VM stop command
+- Add VM start command
+- Add VM deploy command
+- Add VM reboot command
+
 ## 0.1.0
 
 - Initial pack integration